op.compute => anyhow::Result

+++ crates/common/src/types.rs
@@ -3,4 +3,4 @@
 
 pub type JobID = String;
 pub type ErrorMessage = String;
-pub type CreditPoints = f32;
+pub type CreditPoints = f64;
